Fix/Workaround: Round when normalizing
Currently only a workaround, when vectors can't be normalized enough.

+++ src/main/java/de/ecconia/java/opentung/math/Quaternion.java
@@ -103,9 +103,9 @@ public class Quaternion
 		return m;
 	}
 	
-	public float getLength()
+	public double getLength()
 	{
-		return (float) Math.sqrt(a * a + v.lengthSqared());
+		return Math.sqrt(a * a + v.lengthSqared());
 	}
 	
 	@Override
